Jormun: add warning beta api (+ little refactor)
